# platform/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TerraGenius AI API",
    "description": "API for generating and managing infrastructure-as-code.\n\n## Authentication Flow\n\n1. User signs up or logs in using Firebase Authentication (email/password, Google, etc.) via the frontend.\n2. Frontend obtains a Firebase ID token.\n3. Frontend POSTs the ID token to `/auth/token`.\n4. Backend verifies the Firebase token and returns a custom JWT (`accessToken`).\n5. All protected API calls require `Authorization: Bearer <accessToken>`.\n6. The backend validates the JWT and attaches user context.\n\n**Public endpoints:** `/auth/token`, `/status`, `/health`, `/signup`, `/login`, `/reset-password`\n**Protected endpoints:** All `/api/` routes, `/generate`, `/deployments`, `/billing`, `/audit-logs`, etc.\n\n**Future Ready:** To swap auth providers (e.g. Auth0, Cognito), only the `/auth/token` logic and token validation need to change. The rest of the API and security model remains the same.\n\n**Testing:** Use the provided `test_auth_flow.js` script to verify the full login → token exchange → protected API flow.\n\n**Security:** Never send Firebase tokens directly to protected endpoints. Always use the custom JWT returned from `/auth/token`.\n",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Mock API Server"
    }
  ],
  "paths": {
...
    "/auth/token": {
      "post": {
        "summary": "Exchange Firebase token for custom JWT",
        "description": "Accepts a Firebase ID token and returns a custom JWT for use with protected API endpoints.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firebaseToken": {
                    "type": "string",
                    "example": "eyJhbGciOi..."
                  }
                },
                "required": ["firebaseToken"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JWT token returned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /auth/token endpoint"
      }
    },
...
  }
}